Change RC input and drive output for serial drive control

+++ arduino/chomp/chump_targeting.h
@@ -11,7 +11,7 @@ struct Object
   Object() : Distance(0xFFFF), Left_edge(0.0), Right_edge(16.0) { }
 };
 
-float pidSteer(unsigned int num_detections, Detection* detections);
+int16_t pidSteer(unsigned int num_detections, Detection* detections);
 // float error_derivative(float errors*);
 // Detection GetMinDetection(unsigned int num_detections, Detection* detections);
 // Object CallNearestObj(unsigned int num_detections, Detection* detections);
